+++ src/main/java/com/dassda/repository/CommentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
 
-    Long countByDiaryId(Long diaryId);
+    Integer countByDiaryId(Long diaryId);
 }
