+++ src/main/java/com/dassda/repository/LikesRepository.java
@@ -4,5 +4,5 @@ import com.dassda.entity.Likes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikesRepository extends JpaRepository<Likes, Long> {
-    Long countByDiaryId(Long diaryId);
+    Integer countByDiaryId(Long diaryId);
 }
